+++ zrml/rikiddo/src/types.rs
@@ -28,6 +28,18 @@ pub enum Timespan {
     Weeks(u16),
 }
 
+impl Timespan {
+    pub fn to_seconds(&self) -> u32 {
+        match *self {
+            Timespan::Seconds(d) => d,
+            Timespan::Minutes(d) => d * 60,
+            Timespan::Hours(d) => d * 60 * 60,
+            Timespan::Days(d) => u32::from(d) * 60 * 60 * 24,
+            Timespan::Weeks(d) => u32::from(d) * 60 * 60 * 24 * 7,
+        }
+    }
+}
+
 #[derive(Clone, Debug, Decode, Encode, Eq, PartialEq)]
 pub struct RikiddoSigmoidMV<FI: Fixed, FE: RikiddoFee<FI>, MA: MarketAverage<FI>> {
     pub fees: FE,
